move login btn to layout
